unipulse-ai/backend/requirements.txt
fastapi
uvicorn
requests
vaderSentiment
transformers
sqlalchemy
psycopg2-binary
apscheduler
python-dotenv
torch
pandas